refactor(report builder): split logic between static report and gui
