+++ lib/gui/tool-runner/index.js
@@ -8,7 +8,7 @@ const Promise = require('bluebird');
 
 const Runner = require('./runner');
 const subscribeOnToolEvents = require('./report-subscriber');
-const ReportBuilderSqlite = require('../../report-builder/report-builder-sqlite');
+const GuiReportBuilder = require('../../report-builder/gui');
 const EventSource = require('../event-source');
 const utils = require('../../server-utils');
 const {findNode} = require('../../../lib/static/modules/utils');
@@ -58,7 +58,7 @@ module.exports = class ToolRunner {
     async initialize() {
         await mergeDatabasesForReuse(this._reportPath);
 
-        this._reportBuilder = ReportBuilderSqlite.create(this._hermione, this._pluginConfig, {reuse: true});
+        this._reportBuilder = GuiReportBuilder.create(this._hermione, this._pluginConfig, {reuse: true});
         await this._reportBuilder.init();
         await this._reportBuilder.saveStaticFiles();
 
@@ -114,8 +114,7 @@ module.exports = class ToolRunner {
 
     async _fillTestsTree() {
         const {autoRun} = this._guiOpts;
-
-        this._tree = Object.assign(this._reportBuilder.getResult(), {gui: true, autoRun});
+        this._tree = {...this._reportBuilder.getResult(), gui: true, autoRun};
 
         const {suites, browsers} = await this._applyReuseData(this._tree.suites);
