+++ lib/plugin-adapter.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const parseConfig = require('./config');
-const ReportBuilderSqlite = require('./report-builder/report-builder-sqlite');
+const StaticReportBuilder = require('./report-builder/static');
 const utils = require('./server-utils');
 const cliCommands = require('./cli-commands');
 const PluginApi = require('./plugin-api');
@@ -45,21 +45,21 @@ module.exports = class PluginAdapter {
         return this;
     }
 
-    async _createReportBuilderSqlite(prepareData) {
-        const reportBuilderSqlite = ReportBuilderSqlite.create(this._hermione, this._config);
+    async _createStaticReportBuilder(prepareData) {
+        const staticReportBuilder = StaticReportBuilder.create(this._hermione, this._config);
 
-        await reportBuilderSqlite.init();
+        await staticReportBuilder.init();
 
         return Promise
             .all([
-                reportBuilderSqlite.saveStaticFiles(),
-                prepareData(this._hermione, reportBuilderSqlite, this._config)
+                staticReportBuilder.saveStaticFiles(),
+                prepareData(this._hermione, staticReportBuilder, this._config)
             ])
-            .then(() => reportBuilderSqlite.finalize());
+            .then(() => staticReportBuilder.finalize());
     }
 
     async _run(prepareData) {
-        const generateReport = this._createReportBuilderSqlite(prepareData);
+        const generateReport = this._createStaticReportBuilder(prepareData);
 
         this._hermione.on(this._hermione.events.RUNNER_END, () =>
             generateReport.then(() => utils.logPathToHtmlReport(this._config)).catch(utils.logError),
