+++ lib/static/modules/reducers/reporter.js
@@ -87,7 +87,7 @@ export default withBrowserStorage(reducer);
 function reducer(state = getInitialState(compiledData), action) {
     switch (action.type) {
         case actionNames.VIEW_INITIAL: {
-            return getInitialState(action.payload);
+            return getInitialState({...compiledData, ...action.payload});
         }
         case actionNames.RUN_ALL_TESTS: {
             const suites = clone(state.suites);
@@ -317,6 +317,7 @@ function createTestResultsFromDb(state, action) {
     const formattedBrowsers = formatBrowsersDataFromDb(browsersRows);
     const browseres = map(formattedBrowsers, 'id');
     const {suites, suitesStats} = formattedSuites;
+
     const {failed, passed, retries, skipped, total, perBrowser} = suitesStats;
     const suiteIds = {
         all: getSuiteIds(suites).sort(),
